edited BookDetail.css
