+++ client/src/components/RatingCreate/RatingCreate.css
@@ -1,6 +1,6 @@
 .rating-container{
   display: flex;
-  justify-content: center;
+  justify-content: left;
   align-items: center;
 }
