+++ client/src/containers/MainContainer.jsx
@@ -1,6 +1,6 @@
 import { Switch, Route } from "react-router-dom";
 import Books from "../screens/Books/Books";
-import BookDetail from "../screens/BookDetail";
+import BookDetail from "../screens/BookDetail/BookDetail";
 import BookCreate from "../screens/BookCreate";
 import BookEdit from "../screens/BookEdit";
