+++ client/src/screens/Books/Books.css
@@ -24,7 +24,7 @@ select {
   font-size: 18px;
 }
 
-img { 
+.book-image { 
   width: 200px;
 }
 h3 {
